feat: move test to test_forms
