+++ backend/battling/tests/test_forms.py
@@ -1,4 +1,9 @@
+from unittest.mock import patch
+
+from model_bakery import baker
+
 from battling.forms import CreateTeamForm
+from battling.models import PokemonTeam
 from common.utils.tests import TestCaseUtils
 
 
@@ -105,3 +110,69 @@ class CreateTeamFormTest(TestCaseUtils):
         )
 
         self.assertFalse(form.is_valid())
+
+
+class PokemonApiIntegrationTest(TestCaseUtils):
+    def setUp(self):
+        super().setUp()
+        self.opponent = baker.make("users.User")
+        self.battle = baker.make("battling.Battle", creator=self.user, opponent=self.opponent)
+        self.team = baker.make("battling.Team", battle=self.battle, trainer=self.user)
+
+    @patch("pokemon.helpers.get_pokemon_from_api")
+    def test_if_pokemon_api_integration_returns_pokemon_data(self, mock_get_pokemon):
+        def side_effect_func(pokemon_name):
+            fake_json = 1
+            if pokemon_name == "pikachu":
+                fake_json = {
+                    "defense": 40,
+                    "attack": 55,
+                    "hp": 35,
+                    "name": "pikachu",
+                    "img_url": "https://raw.githubusercontent.com"
+                    "/PokeAPI/sprites/master/sprites/pokemon/25.png",
+                    "pokemon_id": 25,
+                }
+            elif pokemon_name == "pidgey":
+                fake_json = {
+                    "defense": 50,
+                    "attack": 25,
+                    "hp": 15,
+                    "name": "pidgey",
+                    "img_url": "https://raw.githubusercontent.com"
+                    "/PokeAPI/sprites/master/sprites/pokemon/25.png",
+                    "pokemon_id": 15,
+                }
+            elif pokemon_name == "bulbasaur":
+                fake_json = {
+                    "defense": 30,
+                    "attack": 40,
+                    "hp": 20,
+                    "name": "bulbasaur",
+                    "img_url": "https://raw.githubusercontent.com"
+                    "/PokeAPI/sprites/master/sprites/pokemon/25.png",
+                    "pokemon_id": 10,
+                }
+            return fake_json
+
+        mock_get_pokemon.side_effect = side_effect_func
+
+        pokemon_data = {
+            "pokemon_1": "pikachu",
+            "pokemon_1_position": 1,
+            "pokemon_2": "pidgey",
+            "pokemon_2_position": 2,
+            "pokemon_3": "bulbasaur",
+            "pokemon_3_position": 3,
+        }
+
+        form = CreateTeamForm(data=pokemon_data)
+        form.instance = self.team
+
+        if form.is_valid():
+            response = form.save()
+            pokemon_team = PokemonTeam.objects.filter(team=response)
+
+        self.assertEqual(pokemon_team[0].pokemon.name, pokemon_data["pokemon_1"])
+        self.assertEqual(pokemon_team[1].pokemon.name, pokemon_data["pokemon_2"])
+        self.assertEqual(pokemon_team[2].pokemon.name, pokemon_data["pokemon_3"])
